+++ code/modules/locations/shuttles/shuttle_window.dm
@@ -13,6 +13,9 @@
 	if (istype(W, /obj/item/weapon/grab) && get_dist(src,user)<2)
 		var/obj/item/weapon/grab/G = W
 		if (istype(G.affecting, /mob/living))
+			if(user.has_trait(TRAIT_PACIFISM))
+				to_chat(user, "<span class='notice'>You don't want to hurt anyone!</span>")
+				return
 			var/mob/living/M = G.affecting
 			var/state = G.state
 			qdel(W)	//gotta delete it here because if window breaks, it won't get deleted
